Rename option package to prelude
