+++ prelude/option_test.go
@@ -1,4 +1,4 @@
-package option
+package prelude
 
 import (
 	"testing"
